requirements.txt
# SentinelStream Dependencies
kafka-python>=2.0.2
pandas>=1.5.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
streamlit>=1.30.0
